pyproject.toml
[tool.poetry]
name  = "iot-anomaly"
version = "0.1.0"
description = "Anomaly detection for sensor data"
authors = ["Moses Owhonda <owhondamoses7@gmail.com>"]

packages = [
  { include = "iot_anomaly", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pydantic = "~1.10.13"
griffe = "~0.30.1"
anyio = "~3.7.1"
joblib = "^1.3.2"
numpy = "^2.1.0"
pandas = "^2.2.0"
scikit-learn = "^1.6.1"
imbalanced-learn = "^0.13.0"
ipykernel="^6.29.0"
optuna = "^4.4.0"
tqdm = "^4.66.1"
prefect = "^2.10.16"
mlflow = "^3.1.1"
tensorflow  = "^2.19.0"
flask = "^2.2.5"
gunicorn = "^20.1.0"
gcsfs = "^2025.7.0"
google-cloud-storage = "^2.14.0"
scipy = "^1.11.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
ruff = "^0.5.0"
black = "^24.4.2"
pre-commit = "^3.7.1"
yamllint = "^1.37.1"

[tool.ruff.lint]
ignore = ["E402"]

[build-system]
requires = ["poetry-core"]
build-backend  = "poetry.core.masonry.api"
